+++ src/HyperRenderer.js
@@ -16,33 +16,16 @@ export default class HyperRenderer {
     }
   }
 
-  project([xo, yo, zo, wo]) {
-    const [x, y, z, w] = this.rotatePoint([xo, yo, zo, wo])
+  _directProject([x, y, z, w]) {
     const zoom = 1 + (w * this.fov) / this.w
     return [x / zoom, y / zoom, z / zoom]
   }
 
-  sliceProject(p1, p2) {
-    const [x1, y1, z1, w1] = p1
-    const [x2, y2, z2, w2] = p2
-    // const [x1, y1, z1, w1] = this.rotatePoint(p1)
-    // const [x2, y2, z2, w2] = this.rotatePoint(p2)
-    if ((this.wSlice - w1) * (this.wSlice - w2) >= 0) {
-      return
-    }
-    const a = (this.wSlice - w1) / (w2 - w1)
-    const x = x1 + a * (x2 - x1)
-    const y = y1 + a * (y2 - y1)
-    const z = z1 + a * (z2 - z1)
-
-    return this.project([x, y, z, this.wSlice])
-    // const zoom = 1 + (this.wSlice * this.fov) / this.w
-    // return [x / zoom, y / zoom, z / zoom]
+  project([xo, yo, zo, wo]) {
+    return this._directProject(this.rotatePoint([xo, yo, zo, wo]))
   }
 
-  slice(p1, p2) {
-    const [x1, y1, z1, w1] = this.rotatePoint(p1)
-    const [x2, y2, z2, w2] = this.rotatePoint(p2)
+  _directSlice([x1, y1, z1, w1], [x2, y2, z2, w2]) {
     if ((this.wSlice - w1) * (this.wSlice - w2) >= 0) {
       return
     }
@@ -52,6 +35,10 @@ export default class HyperRenderer {
     const z = z1 + a * (z2 - z1)
 
     return [x, y, z]
+  }
+
+  slice(p1, p2) {
+    return this._directSlice(this.rotatePoint(p1), this.rotatePoint(p2))
   }
 
   shiftSlice(delta, min, max) {
@@ -102,4 +89,6 @@ export default class HyperRenderer {
     w = w * czw - t * szw
     return [x, y, z, w]
   }
+
+  prepare() {}
 }