+++ src/colorGenerators.js
@@ -16,18 +16,20 @@ export const cellColors =
 export const faceColors =
   ({ colors }) =>
   ({ cell, face, type }) =>
-    colors[((type === 'faces' ? face : cell) || 0) % colors.length]
+    colors[((face !== undefined ? face : cell) || 0) % colors.length]
 
 export const wDepthColors = ({ shape, colors }) => {
   const ws = shape.vertices.map(([, , , w]) => w)
   const wmin = Math.min(...ws)
   const wmax = Math.max(...ws)
   return ({ vertex }) =>
-    new Color().lerpColors(
-      colors[0],
-      colors[1],
-      (shape.vertices[vertex][3] - wmin) / (wmax - wmin)
-    )
+    vertex !== undefined
+      ? new Color().lerpColors(
+          colors[0],
+          colors[1 % colors.length],
+          (shape.vertices[vertex][3] - wmin) / (wmax - wmin)
+        )
+      : colors[0]
 }
 
 export const depthColors = ({ shape, colors }) => {
@@ -45,25 +47,28 @@ export const depthColors = ({ shape, colors }) => {
   const wmax = Math.max(...ws)
 
   return ({ vertex }) => {
+    if (vertex === undefined) {
+      return colors[0]
+    }
     const [x, y, z, w] = shape.vertices[vertex]
     const xColor = new Color().lerpColors(
       colors[0],
-      colors[1],
+      colors[1 % colors.length],
       (x - xmin) / (xmax - xmin)
     )
     const yColor = new Color().lerpColors(
-      colors[2],
-      colors[3],
+      colors[2 % colors.length],
+      colors[3 % colors.length],
       (y - ymin) / (ymax - ymin)
     )
     const zColor = new Color().lerpColors(
-      colors[4],
-      colors[5],
+      colors[4 % colors.length],
+      colors[5 % colors.length],
       (z - zmin) / (zmax - zmin)
     )
     const wColor = new Color().lerpColors(
-      colors[6],
-      colors[7],
+      colors[6 % colors.length],
+      colors[7 % colors.length],
       (w - wmin) / (wmax - wmin)
     )
     return xColor.lerp(yColor, 0.5).lerp(zColor, 0.5).lerp(wColor, 0.5)