+++ src/index.js
@@ -1,5 +1,6 @@
 import HyperMesh from './HyperMesh'
 import HyperRenderer from './HyperRenderer'
+import HyperRendererCached from './HyperRendererCached'
 import HyperSlice from './HyperSlice'
 import * as shapes from './shapes'
 import {
@@ -14,6 +15,7 @@ import { normalizeShape } from './helpers'
 export {
   HyperMesh,
   HyperRenderer,
+  HyperRendererCached,
   HyperSlice,
   normalizeShape,
   uniformColors,